requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
pymongo[srv]>=3.12
certifi
python-dotenv
pyyaml
xgboost